fix : 로직 추가 및 변경된 코드로 인한 테스트 코드 전반적으로 수정

+++ src/main/java/com/example/delivery/domain/menu/dto/request/MenuCreateDto.java
@@ -8,9 +8,14 @@ import lombok.Getter;
 @Builder
 public class MenuCreateDto {
 
-  @NotNull private Long storeId;
-  @NotNull private String menuName;
-  @NotNull private Integer price;
-  @NotNull private String description;
-  @NotNull private String photo;
+  @NotNull
+  private Long storeId;
+  @NotNull
+  private String menuName;
+  @NotNull
+  private Long price;
+  @NotNull
+  private String description;
+  @NotNull
+  private String photo;
 }
